+++ app/local/local_route.py
@@ -13,13 +13,4 @@ def geocode():
 
     return local_service.geocode(address)
 
-@local_bp.route("/test", methods=['GET']) #임시라우트네이밍
-def get_weather_by_address():
-    address = request.args.get("address")
-    if not address:
-        raise BadRequestException("주소지는 필수 값입니다.")
-
-    # LocalService에서 새로 추가된 메서드 호출
-    return local_service.get_weather_by_address(address)
-
